requirements.txt
streamlit
pandas
matplotlib
plotly
scipy
numpy
google-generativeai